auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
python-jose
email-validator
pydantic[email]
cryptography
requests
PyJWT>=2.8.0
bcrypt==3.2.2  # ✅ fix version to avoid AttributeError
passlib[bcrypt]==1.7.4
python-dotenv
pymysql